+++ src/assets/mapStyles/mapStyles.js
@@ -3,23 +3,23 @@ export default {
         version: 8,
         sources: {
             basemap: {
-                type: "vector",
+                type: 'vector',
                 // This URL will pull the basemap GeoJson from S3 so that no local tile server is required.
                 // NOTE: this location is under development and may change. If you suddenly lose parts of the map, it
                 // may be that we have moved the tiles to a new location on S3.
-                "tiles": ["https://d38anyyapxci3p.cloudfront.net/baseTiles_3/{z}/{x}/{y}.pbf"]
-                // url: "http://localhost:8086/data/basemap.json"
+                'tiles': ['https://d38anyyapxci3p.cloudfront.net/baseTiles_3/{z}/{x}/{y}.pbf']
+                // url: 'http://localhost:8086/data/basemap.json'
                 // The above URL is an example for using a local mbtiles file and a tile server. See the readme for more
                 // information: https://github.com/usgs-makerspace/wbeep-viz#start-run-the-tile-server
             },
             HRU: {
-                type: "vector",
+                type: 'vector',
                 // This URL will pull the Hydrological Response Units(HRUs) GeoJson from S3 so that no local tile server
                 // is required.
-                // NOTE: this location is under development and may change. If you suddenly your HRUs from the map, it
+                // NOTE: this location is under development and may change. If you suddenly lose your HRUs from the map, it
                 // may be that we have moved the HRU tiles to a new location on S3.
-                "tiles": ["https://d38anyyapxci3p.cloudfront.net/tileTemp_3/{z}/{x}/{y}.pbf"]
-                // url: "http://localhost:8085/data/new2.json"
+                'tiles': ['https://d38anyyapxci3p.cloudfront.net/testTiles_5/{z}/{x}/{y}.pbf']
+                // url: 'http://localhost:8085/data/new2.json'
                 // The above URL is an example for using a local mbtiles file and a tile server. See the readme for more
                 // information: https://github.com/usgs-makerspace/wbeep-viz#start-run-the-tile-server
             }
@@ -34,7 +34,7 @@ export default {
                     'visibility': 'visible'
                 },
                 'paint': {
-                    'background-color': 'rgba(202, 210, 211, 1)'
+                    'background-color': 'rgba(233, 245, 245, 0.5)'
                 },
                 'showButton': true
             },
@@ -52,43 +52,29 @@ export default {
                 'showButton': true
             },
             {
-                'id': 'Neighboring Countries',
-                'type': 'fill',
-                'source': 'basemap',
-                'source-layer': 'neighboringcountry',
-                'layout': {
-                    'visibility': 'visible'
-                },
-                'paint': {
-                    'fill-color': 'rgba(246, 246, 244, 1)'
-                },
-                'showButton': true
-            },
-            {
                 'id': 'HRUS Fill Colors',
                 'type': 'fill',
                 'source': 'HRU',
-                'source-layer': 'no_simp_prec5',
+                'source-layer': 'hrus',
                 'layout': {
                     'visibility': 'visible'
                 },
-                'generateId': true,
                 'paint': {
                     'fill-color': {
                         'property': 'SoilMoisture',
                         'type': 'categorical',
                         'stops': [
-                            ["",'#000000'],
-                            ['very low','#CC4C02'],
-                            ['low', '#EDAA5F'],
-                            ['average','#FED98E'],
-                            ['high','#A7B9D7'],
                             ['very high','#144873'],
+                            ['high','#A7B9D7'],
+                            ['average','#FED98E'],
+                            ['low', '#EDAA5F'],
+                            ['very low','#CC4C02'],
+                            ["",'#000000'],
                         ]
                     },
                     'fill-opacity': ['case',
                         ['boolean', ['feature-state', 'hover'], false],
-                        0.5,
+                        0.1,
                         1
                     ]
                 },
@@ -99,7 +85,7 @@ export default {
                 'id': 'HRUS Outlines',
                 'type': 'line',
                 'source': 'HRU',
-                'source-layer': 'no_simp_prec5',
+                'source-layer': 'hrus',
                 'layout': {
                     'visibility': 'visible'
                 },
@@ -117,6 +103,19 @@ export default {
                         ]
                     },
                     'line-width': 1
+                },
+                'showButton': true
+            },
+            {
+                'id': 'Neighboring Countries',
+                'type': 'fill',
+                'source': 'basemap',
+                'source-layer': 'neighboringcountry',
+                'layout': {
+                    'visibility': 'visible'
+                },
+                'paint': {
+                    'fill-color': 'rgba(237, 236, 232, 1)'
                 },
                 'showButton': true
             },
@@ -145,7 +144,7 @@ export default {
                     'visibility': 'visible'
                 },
                 'paint': {
-                    'line-color': 'rgba(115, 255, 255, 1)'
+                    'line-color': 'rgba(218, 234, 240, 1)'
                 },
                 'showButton': true
             },
@@ -158,7 +157,7 @@ export default {
                     'visibility': 'visible'
                 },
                 'paint': {
-                    'line-color': 'rgba(115, 255, 255, 1)',
+                    'line-color': 'rgba(218, 234, 240, 1)',
                     'line-dasharray': [
                         2,
                         1.5
@@ -192,23 +191,23 @@ export default {
                     'text-font': [
                         'Roboto Regular'
                     ],
-                    "text-size": 12,
-                    "symbol-placement": "point",
-                    "text-line-height": 1.2,
-                    "text-justify": "center",
-                    "text-anchor": "bottom",
-                    "text-offset": [
+                    'text-size': 12,
+                    'symbol-placement': 'point',
+                    'text-line-height': 1.2,
+                    'text-justify': 'center',
+                    'text-anchor': 'bottom',
+                    'text-offset': [
                         0,
                         -0.5
                     ]
                 },
-                "paint": {
-                    "text-color": "rgba(255,255,255, 1)",
-                    "text-halo-width": 1,
-                    "text-halo-blur": 1,
-                    "text-halo-color": "rgba(0, 0, 0, 0.5)",
+                'paint': {
+                    'text-color': 'rgba(255,255,255, 1)',
+                    'text-halo-width': 1,
+                    'text-halo-blur': 1,
+                    'text-halo-color': 'rgba(0, 0, 0, 0.5)',
                 },
-                "showButton": true
+                'showButton': true
             }
         ]
     }